Changed -output_xhtml- to -output_folder-
The label for the variable and key in the cfg file did not align with what the setting was.

Note that the settings_file.cfg currently reflects Larry's computer and the settings will need to be made manually initially for other computers.

Should probably figure out a way to put the cfg file somewhere else otherwise it will get overwritten by the developers machine all the time.